+++ CHANGELOG.md
@@ -1,15 +1,23 @@
 ## CHANGELOG
 
-## v1.1.1 [2024-01-14]
+## v1.2.0 [2024-01-19]
 ### New Features
+- Generate keys from known mapping
+
+### Other Changes
+- Moved random key to 'Generate Key' frame
+- Key required for encrypting files
+
+## v1.1.1 [2024-01-14]
+### Improvements
 - Error pop-ups for invalid key inputs
 
 ### Bug Fixes
 - Fixed error when key input is not an integer
 
 ## v1.1.0 [2024-01-13]
-### New Features
+### Improvements
 - Updated GUI
 
 ### Other Changes
-- Improved readability
+- Improved code readability